feat(layout): pages & components
