+++ src/app.html
@@ -10,7 +10,7 @@
     <meta name="viewport" content="width=device-width" />
     %sveltekit.head%
   </head>
-  <body data-sveltekit-preload-data="hover" data-theme="rocket">
-    <div style="display: contents">%sveltekit.body%</div>
+  <body>
+    <div style="display: contents" class="h-full overflow-hidden">%sveltekit.body%</div>
   </body>
 </html>
