+++ src/app.postcss
@@ -1,7 +1,6 @@
-/*place global styles here */
 html,
 body {
-  @apply h-full;
+  @apply h-full overflow-hidden;
 }
 
 @font-face {
