+++ src/lib/configuration.ts
@@ -3,7 +3,8 @@ export const Configuration = {
     Name: 'NATO Phonetic Alphabet'
   },
   Github: {
-    Title: 'Ealenn/NATO-phonetic-alphabet',
-    Link: 'https://github.com/Ealenn/NATO-phonetic-alphabet'
+    User: 'Ealenn',
+    Repository: 'NATO-phonetic-alphabet',
+    Url: 'https://github.com/Ealenn/NATO-phonetic-alphabet'
   }
 };
