+++ src/routes/+page.svelte
@@ -1,50 +1,20 @@
 <script lang="ts">
-  import { AppRail, AppRailAnchor, AppRailTile } from '@skeletonlabs/skeleton';
-  import AlphabetTab from './tabs/alphabet.svelte';
-  import TrainingTab from './tabs/training.svelte';
+  import { AlphabetService } from '$lib/alphabet';
   import { Configuration } from '../lib/configuration';
+  import LetterCard from '../components/letter-card.svelte';
 
-  let currentTile = 'alphabet';
+  const alphabetService = new AlphabetService();
+  const alphabet = alphabetService.getAll();
 </script>
 
 <svelte:head>
-  <title
-    >{Configuration.Project.Name} - {currentTile.charAt(0).toUpperCase() +
-      currentTile.slice(1)}</title
-  >
+  <title>{Configuration.Project.Name} - Alphabet</title>
 </svelte:head>
 
-<div class="card !bg-transparent grid grid-cols-[auto_1fr] h-full w-full">
-  <AppRail>
-    <AppRailTile bind:group={currentTile} name="alphabet" value="alphabet" title="alphabet">
-      <svelte:fragment slot="lead">
-        <i class="ti ti-abc" />
-      </svelte:fragment>
-      <span>Alphabet</span>
-    </AppRailTile>
-    <AppRailTile bind:group={currentTile} name="training" value="training" title="training">
-      <svelte:fragment slot="lead">
-        <i class="ti ti-cards" />
-      </svelte:fragment>
-      <span>Training</span>
-    </AppRailTile>
-
-    <svelte:fragment slot="trail">
-      <AppRailAnchor
-        href={Configuration.Github.Link}
-        target="_blank"
-        title={Configuration.Github.Title}
-      >
-        <i class="ti ti-brand-github-filled" />
-      </AppRailAnchor>
-    </svelte:fragment>
-  </AppRail>
-
-  {#if currentTile === 'alphabet'}
-    <AlphabetTab />
-  {/if}
-
-  {#if currentTile === 'training'}
-    <TrainingTab />
-  {/if}
+<div class="flex items-center justify-center h-screen p-5">
+  <div class="container m-auto flex flex-wrap flex-col md:flex-row items-center justify-start">
+    {#each alphabet as letter}
+      <LetterCard {letter} additionalClass="m-1" />
+    {/each}
+  </div>
 </div>
